Adding structure and Hook
