+++ TavalaExtension/Hooks/Order/SubmitOrderHook.cs
@@ -1,0 +1,19 @@
+﻿using DirectScale.Disco.Extension.Hooks;
+using DirectScale.Disco.Extension.Hooks.Orders;
+using System;
+using System.Threading.Tasks;
+
+namespace TavalaExtension.Hooks.Order
+{
+    public class SubmitOrderHook : IHook<SubmitOrderHookRequest, SubmitOrderHookResponse>
+    {
+        public SubmitOrderHook()
+        {
+        }
+        public Task<SubmitOrderHookResponse> Invoke(SubmitOrderHookRequest request, Func<SubmitOrderHookRequest, Task<SubmitOrderHookResponse>> func)
+        {
+            var result = func(request);
+            return result;
+        }
+    }
+}
