--- TavalaExtension/Pages/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-@using TavalaExtension
-@namespace TavalaExtension.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
