+++ TavalaExtension/TavalaExtension.csproj
@@ -8,4 +8,12 @@
     <PackageReference Include="DirectScale.Disco.Extension.Abstractions" Version="1.0.5214" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Controllers\" />
+    <Folder Include="Hooks\Order\" />
+    <Folder Include="Models\" />
+    <Folder Include="Services\" />
+    <Folder Include="Repositories\" />
+  </ItemGroup>
+
 </Project>
