+++ TavalaExtension/Views/Shared/_ValidationScriptsPartial.cshtml
@@ -1,0 +1,2 @@
+﻿<script src="~/lib/jquery-validation/dist/jquery.validate.min.js"></script>
+<script src="~/lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.min.js"></script>
