+++ TavalaExtension/Views/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+@using TavalaExtension
+@namespace TavalaExtension.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
